==> 4_tidy_clindat.Rmd <==
---
title: "4_tidy_clindat"
output: 
  pdf_document:
    keep_md: TRUE
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Generated on: `r Sys.time()`

Import required packages
```{r}
library(tidyverse)
library(readxl)
library(lubridate)
library(eeptools)
library(survival)
library(survminer)
```

```{r}
clinDataFn <- "./input/datasets_220405/KS_CRCLM_2012-2021_first_310_Jennie_completed_20211210_pure_values.xlsx" # KS_CRLM_2012_2025_pat_260_20211212.xlsx" # "CRLM_clin_cohort_210315.xlsx"
patDataFn1 <- "./input/datasets_220405/CRLM_KS_cohort_pathology_Evelina.xlsx" # Evelina: 1-58 # CRLM_KS_cohort_pathology_Evelina (200904) # Lorand (210309)
patDataFn2 <- "./input/datasets_220405/CRLM_KS_cohort_pathology_Media-Yousra.xlsx" # Lorand/Danyil-Media-Yousra: 59-310 # "Lorand"

gpProbeFn <- "./output/gp_annotations_by_probe.csv"
tumorPercentProbeFn <- "./output/regression_by_probe.csv"

tumorPercentTumorFn <- "./output/regression_by_tumor.csv"
gpAnnotationsTumorFn <- "./output/gp_annotations_by_tumor.csv"

clinPatAnnotFn <- "./output/clin_data_annot.csv"

dateEndFu <- as.Date("2022-02-26") #  Next: "2021-03-08" # Initial Evelina's cohort as.Date("2020-10-29") # Current date of end follow up for study
# Obs, check that survival data is updated to the last date for previous (Evelina's) cohort

is_visual_scores = FALSE # TRUE   #     Use visual scores data
is_until_104 = FALSE # TRUE # For comparision annotation vs visual with same n
```

Import data
```{r}
# Modifications of CRLM_clin_cohort.xlsx:
#     -Deleted row 39: repeated case, see XZXZ	2/13/1973
#     -Row 6: date of progression in liver:  2012-11-19 > 11/19/2012

# Doubts: Ids 2 and 6, same date of date of progression and progression in liver
#         Id 11, no date for liver op   
 
# https://readxl.tidyverse.org/articles/cell-and-column-types.html
# Peek at column names in CRLM_clin_cohort.xlsx
(nms <- names(read_excel(clinDataFn, n_max = 0)))

# Columns to import from excel
# [1] "PaRaMet nr...1"                       "PaRaMet nr...2"                       "pad"                                  "dob"                                  "sex"                                 
# [6] "location of primary (left/right)"     "date of liver op"                     "date of death"                        "Cancer Related Death (Y/N/, unclear)" "date of progression/recidive"        
#[11] "date of progression in liver"         "neoadjvant (Y/N)"                     "sync/meta (s/m)"                      "lost for follow up"                   "comment"   

# Read excel file
clinData <- read_excel(clinDataFn, col_types = c("skip", "skip", "text", "date", "text", "text", "date", "date", "text",  "date", "date", "text", "text", "date", "skip"))

# Rename columns with easy and concise names: rename(flights, tail_num = tailnum)
(nms <- names(clinData))
clinData <- rename(clinData, date_birth = "dob", primary_side = "location of primary (left/right)",
                   date_liver_op = "date of liver op", date_death = "date of death", 
                   cancer_death = "Cancer Related Death (Y/N/U, unclear)", date_progression = "date of progression/recidive", 
                   date_progression_liver = "date of progression in liver", neoadjuvant = "neoadjuvant (Y/N)", sync_meta = "sync/meta (s/m)",
                   date_lost_follow_up = "date of lost for follow up")
(nms <- names(clinData))

# Remove wrong rows
#clinData <- clinData %>% filter(!id %in% c(11, NA))

# Filter for individual cohorts
#clinData <- clinData #  %>% filter(id < 59) # Evelina < 59, Lorand-Danyil >= 59
# Remove wrong rows
#clinData <- clinData %>% filter(!id %in% c(61, 70)) # TODO find personnummer id 70

# Reorder columns
clinData <- clinData %>% relocate(sex, .after = pad)
clinData <- clinData %>% relocate(date_progression, .after = date_liver_op)
clinData <- clinData %>% relocate(date_progression_liver, .after = date_progression)
clinData <- clinData %>% relocate(date_lost_follow_up, .after = date_death)
clinData <- clinData %>% relocate(sync_meta, .after = date_lost_follow_up)
clinData <- clinData %>% relocate(neoadjuvant, .after = sync_meta)
(nms <- names(clinData))

# Convert character values as factors or logical
clinData$primary_side <- as.factor(clinData$primary_side) 
clinData$sync_meta <- as.factor(clinData$sync_meta) # S(ynchronous) / M(etachronous)
clinData$sync_meta <- fct_relevel(clinData$sync_meta, "M", "S")

# TODO: Handle cancer_death Unknown
clinData$cancer_death <- as.factor(clinData$cancer_death) # Y(es) / N(o) / U(nknown)
clinData$cancer_death <- fct_relevel(clinData$cancer_death, "Y", "N")

clinData$neoadjuvant <- as.factor(clinData$neoadjuvant) # Y(es) / N(o)
clinData$neoadjuvant <- fct_relevel(clinData$neoadjuvant, "Y", "N")

# Generate derived columns: time_follow_up, status_follow_up (1 - death observed, 0 - survival time censored)
# MOST IMPORTANT FOR SURV ANALYSYS, ALWAYS CHECK
fuDeath <- clinData %>% filter(!is.na(date_death)) %>% mutate(time_fu_os = as.double(difftime(date_death, date_liver_op, units = c("days"))), status_fu_os = 1)
fuAlive <- clinData %>% filter(is.na(date_death) & is.na(date_lost_follow_up)) %>% mutate(time_fu_os = as.double(difftime(dateEndFu, date_liver_op, units = c("days"))), status_fu_os = 0)
fuLostOS <- clinData %>% filter(!is.na(date_lost_follow_up) & is.na(date_death)) %>% mutate(time_fu_os = as.double(difftime(date_lost_follow_up, date_liver_op, units = c("days"))), status_fu_os = 0) # "T7303/2015" lost fu 2015 but dod known 2016
clinSurvDataOS <- rbind(fuDeath, fuAlive, fuLostOS)
stopifnot(all_equal(clinData, select(clinSurvDataOS, -time_fu_os, -status_fu_os)))

fuProgress <- clinData %>% filter(!is.na(date_progression_liver)) %>% mutate(time_fu_prog = as.double(difftime(date_progression_liver, date_liver_op, units = c("days"))), status_fu_prog = 1)
fuProgressDeath <- clinData %>% filter(is.na(date_progression_liver) & !is.na(date_death)) %>% mutate(time_fu_prog = as.double(difftime(date_death, date_liver_op, units = c("days"))), status_fu_prog = 1)
fuProgFree <- clinData %>% filter(is.na(date_progression_liver) & is.na(date_death) & is.na(date_lost_follow_up)) %>% mutate(time_fu_prog = as.double(difftime(dateEndFu, date_liver_op, units = c("days"))), status_fu_prog = 0)
fuProgLost <- clinData %>% filter(is.na(date_progression_liver) & !is.na(date_lost_follow_up) & is.na(date_death)) %>% mutate(time_fu_prog = as.double(difftime(date_lost_follow_up, date_liver_op, units = c("days"))), status_fu_prog = 0) # "T7303/2015" lost fu 2015 but dod known 2016
clinSurvDataPFS <- rbind(fuProgress, fuProgressDeath, fuProgFree, fuProgLost)
stopifnot(all_equal(clinData, select(clinSurvDataPFS, -time_fu_prog, -status_fu_prog)))

clinSurvData <- merge(clinSurvDataOS, select(clinSurvDataPFS, pad, time_fu_prog, status_fu_prog))
stopifnot(all_equal(clinData, select(clinSurvData, -time_fu_os, -status_fu_os, -time_fu_prog, -status_fu_prog)))

#' Calculate age
#' 
#' By default, calculates the typical "age in years", with a
#' \code{floor} applied so that you are, e.g., 5 years old from
#' 5th birthday through the day before your 6th birthday. Set
#' \code{floor = FALSE} to return decimal ages, and change \code{units}
#' for units other than years.
#' @param dob date-of-birth, the day to start calculating age.
#' @param age.day the date on which age is to be calculated.
#' @param units unit to measure age in. Defaults to \code{"years"}. Passed to \link{\code{duration}}.
#' @param floor boolean for whether or not to floor the result. Defaults to \code{TRUE}.
#' @return Age in \code{units}. Will be an integer if \code{floor = TRUE}.
#' @examples
#' my.dob <- as.Date('1983-10-20')
#' age(my.dob)
#' age(my.dob, units = "minutes")
#' age(my.dob, floor = FALSE)
age <- function(dob, age.day = today(), units = "years", floor = TRUE) {
    calc.age = interval(dob, age.day) / duration(num = 1, units = units)
    if (floor) return(as.integer(floor(calc.age)))
    return(calc.age)
}
clinSurvData <- clinSurvData %>% mutate(age = age(date_birth, date_liver_op))
clinSurvData <- clinSurvData %>% relocate(age, .after = pad)

# Dichotomize age
hist(clinSurvData$age)
clinSurvData <- clinSurvData %>% mutate(age_group = ifelse(age >=60, "60+", "<60"))
clinSurvData <- clinSurvData %>% relocate(age_group, .after = age)
```

Merge clinical with pathology report data
```{r}
# Peek at column names in CRLM_KS_cohort_pathology
(nms <- names(read_excel(patDataFn1, n_max = 0)))
(nms <- names(read_excel(patDataFn2, n_max = 0)))

# Read pathology report data
# Evelina's dataset
patDataOrig1 <- read_excel(patDataFn1, col_types = c(
  "skip", "numeric", "text", "numeric", "text",  rep("skip", 8),
  rep(c("numeric", "skip", "numeric", "numeric"), 16)
  ))
# Lorand/Danyil-Media-Yousra dataset
patDataOrig2 <- read_excel(patDataFn2, col_types = c(
  "skip", "numeric", "text", "numeric", "text",  rep("skip", 8),
  rep(c("numeric", "skip", "numeric", "numeric"), 16)
  ))

# Merge Evelina's and Lorand's pathology datasets
patDataOrig <- rbind(patDataOrig1, patDataOrig2)

patDataTemp <-patDataOrig
# Complete missing tumor % orig with review
# https://www.tidyverse.org/blog/2019/06/rlang-0-4-0/#other-simple-tidy-evaluation-patterns
# Also https://stackoverflow.com/questions/29678435/how-to-pass-dynamic-column-names-in-dplyr-into-custom-function
tumor_letters <- c('a' , 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p')
for (a_letter in tumor_letters) {
  percent_tumor_col_name <- str_glue('tumor_{a_letter}_percent_tumor (%)')
  percent_tumor_orig_col_name <- str_glue('tumor_{a_letter}_percent_tumor_orig (%)')
  percent_tumor_review_col_name <- str_glue('tumor_{a_letter}_percent_tumor_review (%)')
  
  patDataTemp <- patDataTemp %>% mutate( {{percent_tumor_col_name}} := ifelse(is.na(.data[[percent_tumor_orig_col_name]]), .data[[percent_tumor_review_col_name]], .data[[percent_tumor_orig_col_name]]))
  patDataTemp <- patDataTemp %>% relocate({{percent_tumor_col_name}}, .before = {{percent_tumor_orig_col_name}})
}

# Derive max tumour diameter per probe
patDataTemp <- patDataTemp %>% rowwise() %>% mutate(max_diameter = max(c_across(contains("largest_diameter (cm)")), na.rm = TRUE))

# Derive sum of tumor diameters per probe
patDataTemp <- patDataTemp %>% rowwise() %>% mutate(sum_diameter = sum(c_across(contains("largest_diameter (cm)")), na.rm = TRUE))

# Derive mean regression per probe
patDataTemp <- patDataTemp %>% rowwise() %>% mutate(mean_tumor_percent_report = mean(c_across(contains("percent_tumor (%)")), na.rm = TRUE))
patDataTemp <-patDataTemp %>% mutate(mean_tumor_percent_report = ifelse(is.nan(mean_tumor_percent_report), NA, mean_tumor_percent_report) )

patDataTemp$mean_tumor_percent_report <- as.numeric(patDataTemp$mean_tumor_percent_report)

# Derive sum regression score per probe (sum of % tumor x diameter)
# TODO: sum of NA (or a product of it) is 0. We need to garantee every diameter is matched with percent tumour. Add integrity check code
ids_incomplete_tumor_report <- c()
#ids_incomplete_tumor_report_a <- c()
for(a_letter in tumor_letters) {
  largest_diameter_col_name <- str_glue('tumor_{a_letter}_largest_diameter (cm)')
  percent_tumor_col_name <- str_glue('tumor_{a_letter}_percent_tumor (%)')
  
  # We know a tumor exists is diameter is not NA, obtain them
  paired_columns <- patDataTemp %>% select(pad, {{largest_diameter_col_name}}, {{percent_tumor_col_name}}) %>% filter(!is.na(.data[[largest_diameter_col_name]]))
  # Obtain if for any existing tumor (largest diameter != NA) percent tumor is missing (NA)
  incomplete_cases <- paired_columns %>% filter(is.na(.data[[percent_tumor_col_name]]))
  
  for(an_id in incomplete_cases$projectnr) {
    ids_incomplete_tumor_report <- c(ids_incomplete_tumor_report, an_id)
    warning(str_glue("Missing tumor percent report for probe id {an_id}, tumor {a_letter} "))
    print("")
  }
}
  
for(a_letter in tumor_letters) {
  percentxdiam_col_name <- str_glue('percentxdiam_{a_letter}') 
  largest_diameter_col_name <- str_glue('tumor_{a_letter}_largest_diameter (cm)')
  percent_tumor_col_name <- str_glue('tumor_{a_letter}_percent_tumor (%)')
  
  patDataTemp <- patDataTemp %>% mutate({{percentxdiam_col_name}} := .data[[largest_diameter_col_name]] * .data[[percent_tumor_col_name]])
  patDataTemp <- patDataTemp %>% relocate({{percentxdiam_col_name}}, .after = {{percent_tumor_col_name}})
}
# Derive sum regression score per probe (sum of % tumor x diameter)
# TODO: sum of NA (or a product of it) is 0. We need to garantee every diameter is matched with percent tumour. Add integrity check code
patDataTemp <- patDataTemp %>% mutate(sum_tumour_score_report = sum(c_across(starts_with("percentxdiam")), na.rm = TRUE))

# Prev UGLY AND BUGGY HACK - We loose complete regressions, fix above. Remove 0 because % tumor_a is NA
patDataTemp <-patDataTemp %>% mutate(sum_tumour_score_report = na_if(sum_tumour_score_report, 0.0))

patDataTemp <- patDataTemp %>% relocate(sum_tumour_score_report, .before = `tumor_a_percent_tumor (%)`)

# Leave out per tumour nodule columns
patData <- patDataTemp %>% select(-contains("largest_diameter"), -contains("tumor_orig"), -contains("tumor_review"), -contains("percent_tumor (%)"), -contains("percentxdiam"))

# Rename columns
patData <- rename(patData, marginal = "min_dist_rr (mm)")
                   
# Convert marginal <1 into 0.5 and make it numeric
patData$marginal <- gsub("<1", "0.5", patData$marginal, fixed = TRUE)
patData$marginal <- as.numeric(patData$marginal)

# Dichotomize num_tumors
table(patData$num_tumors)
hist(patData$num_tumors, breaks = 20)
patData <- patData %>% mutate(num_tumors_group = ifelse(num_tumors >=4, "4+", "<4"))
patData <- patData %>% relocate(num_tumors_group, .after = num_tumors)

# Dichotomize largest tumors diameter
table(patData$max_diameter)
hist(patData$max_diameter)
patData <- patData %>% mutate(max_diameter_group = ifelse(max_diameter >=5, "5+", "<5"))
patData <- patData %>% relocate(max_diameter_group, .after = max_diameter)

# Dichotomize sum tumor diameters
table(patData$sum_diameter)
hist(patData$sum_diameter)
patData <- patData %>% mutate(sum_diameter_group = ifelse(sum_diameter >=9, "9+", "<9")) # Significant already with 7
patData <- patData %>% relocate(sum_diameter_group, .after = sum_diameter)

# Dichotomize marginal with 0 and 1 mm as cut-offs
table(patData$marginal)
hist(patData$marginal, freq = TRUE, breaks = c(0, 1, 5, seq(10, max(patData$marginal, na.rm = TRUE), by = 10)))
patData <- patData %>% mutate(marginal_group0 = ifelse(marginal > 0, ">0", "0"))
patData <- patData %>% relocate(marginal_group0, .after = marginal)

patData <- patData %>% mutate(marginal_group1 = ifelse(marginal >= 1, "1+", "<1"))
patData <- patData %>% relocate(marginal_group1, .after = marginal_group0)

patData <- patData %>% mutate(marginal_group01 = ifelse(marginal == 0, "0", ifelse(marginal >= 1, "1+", "<1") ))
patData <- patData %>% relocate(marginal_group01, .after = marginal_group1)

# Dichotomize mean tumor percent
table(patData$mean_tumor_percent_report)
hist(patData$mean_tumor_percent_report, freq = TRUE, breaks = c(0, 5, seq(10, 100, by = 10)))
patData <- patData %>% mutate(tum_percent_report_group = ifelse(mean_tumor_percent_report < 50, "<50", "50+"))
patData <- patData %>% relocate(tum_percent_report_group, .after = mean_tumor_percent_report)

# Dichotomize sum regression score
table(patData$sum_tumour_score_report)
hist(patData$sum_tumour_score_report)
patData <- patData %>% mutate(sum_tumour_score_report_group = ifelse(sum_tumour_score_report >= 250, "250+", "<250"))
patData <- patData %>% relocate(sum_tumour_score_report_group, .after = sum_tumour_score_report)

# Merge clinical and pathology report data
clinPatData <- merge(clinSurvData, patData, by = "pad")

if(is_visual_scores || is_until_104) { # Limit dataset to Evelina and Danyil
  clinPatData <- clinPatData %>% filter(projectnr < 105)
}
```

Read and process annotation derived GP data
```{r }
gpProbeOrig <- read.csv(gpProbeFn, row.names=NULL)
gpProbeOrig$annotation_types <- fct_relevel(gpProbeOrig$annotation_types, "P", "D","R2", "R")

# Rename columns
gpProbeOrig <- gpProbeOrig %>% rename(id = ids)

# Classify each probe accordibng to predominant GP (4 categories)
gpPredominant4 <- gpProbeOrig %>% group_by(id) %>% slice_max(percent_gp) %>% select(id, annotation_types)
gpPredominant4 <- gpPredominant4 %>% rename(gp_predo4 = annotation_types)
fifty_fifty <- gpPredominant4 %>% filter(duplicated(id)) %>% pull(id)
gpPredominant4 <- gpPredominant4 %>% filter(!(id %in% fifty_fifty))

# Classify each probe accordibng to predominant GP (3 categories)
gPredoDesmoPushing <- gpProbeOrig %>% filter(annotation_types %in% c("D", "P")) %>% select(-length_um)
gPredoRepl12 <- gpProbeOrig %>% filter(annotation_types %in% c("R", "R2"))
gPredoRepl <- gPredoRepl12 %>% group_by(id) %>% summarise(annotation_types = "R", percent_gp = sum(percent_gp))

gpPredominant3 <- rbind(gPredoDesmoPushing, gPredoRepl) %>% group_by(id) %>% slice_max(percent_gp) %>% select(id, annotation_types)
gpPredominant3 <- gpPredominant3 %>% rename(gp_predo3 = annotation_types)
fifty_fifty <- gpPredominant3 %>% filter(duplicated(id)) %>% pull(id)
gpPredominant3 <- gpPredominant3 %>% filter(!(id %in% fifty_fifty))

# Initialize % of replacement and desmoplastic for every tumor, so those with 0 % are captured when filtering
for(an_id in unique(gpProbeOrig$id)) {
  gpProbeOrig <- gpProbeOrig %>% add_row(id = an_id, annotation_types = "R", length_um = 0, percent_gp = 0)
  gpProbeOrig <- gpProbeOrig %>% add_row(id = an_id, annotation_types = "R2", length_um = 0, percent_gp = 0)
  gpProbeOrig <- gpProbeOrig %>% add_row(id = an_id, annotation_types = "D", length_um = 0, percent_gp = 0)
}

# Derive % replacement and desmo per probe
gpProbeReplacement <- gpProbeOrig %>% group_by(id) %>% filter(annotation_types == "R" | annotation_types == "R2") %>% summarise(replacement_percent = sum(percent_gp) )
gpProbeDesmo <- gpProbeOrig %>% group_by(id) %>% filter(annotation_types == "D") %>% summarise(desmo_percent = sum(percent_gp) )

# Derive % replavement 1 vs 2
gpProbeRepl12 <- gpProbeOrig %>% filter(annotation_types == "R" | annotation_types == "R2") %>% group_by(id, annotation_types) %>% summarise(sum_percent_gp = sum(percent_gp))
gpProbeRepl12No0 <- gpProbeRepl12 %>% group_by(id) %>% filter(sum(sum_percent_gp) > 0)
gpProbeRepl12Percent <- gpProbeRepl12No0 %>% group_by(id) %>% mutate(percent_repl = 100 / sum(sum_percent_gp) * sum_percent_gp)
gpRepl2PercentR2 <- gpProbeRepl12Percent %>% filter(annotation_types == "R2")
gpRepl2PercentR2Trim <- gpRepl2PercentR2 %>% select(id, percent_repl)
gpRepl2PercentR2Trim <- gpRepl2PercentR2Trim %>% rename(percent_r2 = percent_repl)

# Dichotomize % replacement
table(gpProbeReplacement$replacement_percent)
hist(gpProbeReplacement$replacement_percent)
gpProbeReplacement <- gpProbeReplacement %>% mutate(replacement_group = ifelse(replacement_percent == 0, "0",
                                                                ifelse(replacement_percent > 0 & replacement_percent < 50, "1-50", "50+")))

# Dichotomize % desmo
table(gpProbeDesmo$desmo_percent)
hist(gpProbeDesmo$desmo_percent)
gpProbeDesmo <- gpProbeDesmo %>% mutate(desmo_group95 = ifelse(desmo_percent >=95, "95+", "<95"))
gpProbeDesmo <- gpProbeDesmo %>% mutate(desmo_group10050 = ifelse(desmo_percent == 100, "100",
                                                                ifelse(desmo_percent >= 50 & desmo_percent < 100, "50-100", "0-50")))
# TODO: refactor discretization code from ifelse to cut
gpProbeDesmo <- gpProbeDesmo %>% mutate(desmo_galjart = cut(gpProbeDesmo$desmo_percent, breaks = c(0, 0.1, 33, 66, 99.9, 100), include.lowest = TRUE))
gpProbeDesmo <- gpProbeDesmo %>% mutate(desmo_ks = cut(gpProbeDesmo$desmo_percent, breaks = c(0, 33, 99.9, 100), include.lowest = TRUE))
gpProbeDesmo <- gpProbeDesmo %>% mutate(desmo_ks2 = cut(gpProbeDesmo$desmo_percent, breaks = c(0, 33, 95, 100), include.lowest = TRUE))

# Dichotomize % replacement2
table(gpRepl2PercentR2Trim$percent_r2)
hist(gpRepl2PercentR2Trim$percent_r2)
gpRepl2PercentR2Trim <- gpRepl2PercentR2Trim %>% mutate(r2_group = ifelse(percent_r2 >=50, "50+", "<50"))
```

Read and process annotation derived regression data
```{r }
tumorPercentProbeOrig <- read.csv(tumorPercentProbeFn, row.names=NULL)

# Rename columns
tumorPercentProbe <- tumorPercentProbeOrig %>% rename(id = ids, tum_percent_annot = avg_percent)

# Dichotomize % tumor annot
table(tumorPercentProbe$tum_percent_annot)
hist(tumorPercentProbe$tum_percent_annot)
tumorPercentProbe <- tumorPercentProbe %>% mutate(tum_percent_annot_group = ifelse(tum_percent_annot < 50, "<50", "50+"))
tumorPercentProbe <- tumorPercentProbe %>% mutate(tum_percent_annot_5 = ifelse(tum_percent_annot <= 5, "5", "5+"))
```

Merge annotation derived regression and GP data with clinPath data
```{r }
clinPatData <- clinPatData %>% rename(id = "projectnr") %>% relocate(id, .after = pad)

clinPatAnnotationData <- merge(clinPatData, gpProbeReplacement, by = "id", all.x = TRUE)
clinPatAnnotationData <- merge(clinPatAnnotationData, tumorPercentProbe, by = "id", all.x = TRUE)
clinPatAnnotationData <- merge(clinPatAnnotationData, gpProbeDesmo, by = "id", all.x = TRUE)
clinPatAnnotationData <- merge(clinPatAnnotationData, gpRepl2PercentR2Trim, by = "id", all.x = TRUE)
clinPatAnnotationData <- merge(clinPatAnnotationData, gpPredominant4, by = "id", all.x = TRUE)
clinPatAnnotationData <- merge(clinPatAnnotationData, gpPredominant3, by = "id", all.x = TRUE)
```

Weighted annotation derived regression
```{r }
# Extract id and tumor diameters 
tumDiameters <- patDataTemp %>% select(projectnr, contains("_largest_diameter"))
tumDiameters <- rename(tumDiameters, id = "projectnr")

tumPercentByTumor <- read.csv(tumorPercentTumorFn)
tumPercentByTumorWide <- tumPercentByTumor %>% pivot_wider(names_from = tumors, values_from = avg_percent, names_prefix = "%_tum_ann_")
tumPercentByTumorWide <- tumPercentByTumorWide %>% rename(id = "ids")
tumPercentByTumorDiams <- merge(tumPercentByTumorWide, tumDiameters, by = "id")

# Check needed columns exist, create otherwise
for(a_letter in tumor_letters) {
  a_percent_tum_ann_col_name <- str_glue("%_tum_ann_{a_letter}")
  if(!(a_percent_tum_ann_col_name %in% colnames(tumPercentByTumorDiams))) {
    tumPercentByTumorDiams <-tumPercentByTumorDiams %>% add_column({{a_percent_tum_ann_col_name}} := NA) 
    message("Added non-existent column {a_percent_tum_ann_col_name} to tumPercentByTumorDiams")
  }
}

# TODO refactor repetitive tidyverse code using https://www.tidyverse.org/blog/2019/06/rlang-0-4-0/
# https://github.com/rstudio/cheatsheets/blob/master/tidyeval.pdf
# Set aside % tumor and largest diameter for easier visual assessment
for(a_letter in tumor_letters) {
  largest_diameter_col_name <- str_glue("tumor_{a_letter}_largest_diameter (cm)")
  percent_tum_ann_col_name <- str_glue("%_tum_ann_{a_letter}")
  tumPercentByTumorDiams <- tumPercentByTumorDiams %>% relocate({{largest_diameter_col_name}}, .after = {{percent_tum_ann_col_name}})
}

# TODO: Following two bits share implementation with by report > refactor into a function
ids_incomplete_tumor_annot <- c()
# ids_incomplete_tumor_annot_a <- c()
for(a_letter in tumor_letters) {
  largest_diameter_col_name <- str_glue('tumor_{a_letter}_largest_diameter (cm)')
  percent_annot_tumor_col_name <- str_glue('%_tum_ann_{a_letter}')
  
  # We know a tumor exists is diameter is not NA, obtain them
  paired_columns_annot <- tumPercentByTumorDiams %>% select(id, {{largest_diameter_col_name}}, {{percent_annot_tumor_col_name}}) %>% filter(!is.na(.data[[largest_diameter_col_name]]))
  # Obtain if for any existing tumor (largest diameter != NA) percent tumor is missing (NA)
  incomplete_cases_annot <- paired_columns_annot %>% filter(is.na(.data[[percent_annot_tumor_col_name]]))
  
  for(an_id in incomplete_cases_annot$id) {
    ids_incomplete_tumor_annot <- c(ids_incomplete_tumor_annot, an_id)
    warning(str_glue("Missing tumor percent from annotation for probe id {an_id}, tumor {a_letter} "))
    print("")
  }
}

for(a_letter in tumor_letters) {
  percentxdiam_annot_col_name <- str_glue('percentxdiam_annot_{a_letter}') 
  largest_diameter_col_name <- str_glue('tumor_{a_letter}_largest_diameter (cm)')
  percent_tumor_annot_col_name <- str_glue('%_tum_ann_{a_letter}')
  
  tumPercentByTumorDiams <- tumPercentByTumorDiams %>% mutate({{percentxdiam_annot_col_name}} := .data[[largest_diameter_col_name]] * .data[[percent_tumor_annot_col_name]])
  tumPercentByTumorDiams <- tumPercentByTumorDiams %>% relocate({{percentxdiam_annot_col_name}}, .after = {{percent_tumor_annot_col_name}})
}

# Derive sum regression score per probe (sum of % tumor x diameter)
# TODO: sum of NA (or a product of it) is 0. We need to garantee every diameter is matched with percent tumour. Add integrity check code
tumPercentByTumorDiams <- tumPercentByTumorDiams %>% rowwise() %>% mutate(sum_tumour_score_annot = sum(c_across(starts_with("percentxdiam_annot")), na.rm = TRUE))

# Try to Manage missing
# Prev UGLY AND BUGGY HACK - We loose complete regressions, fix above
tumPercentByTumorDiams <-tumPercentByTumorDiams %>% mutate(sum_tumour_score_annot = na_if(sum_tumour_score_annot, 0.0))
# tumPercentByTumorDiams <-tumPercentByTumorDiams %>% mutate(sum_tumour_score_annot_old = na_if(sum_tumour_score_annot, 0.0))
# Now, remove values for those probes with incomplete tumor regression in report
#index <- tumPercentByTumorDiams$id %in% ids_incomplete_tumor_annot_a
#tumPercentByTumorDiams$sum_tumour_score_annot[index] <- NA

tumPercentByTumorDiams <- tumPercentByTumorDiams %>% relocate(sum_tumour_score_annot, .after = id)

# Leave out per tumour nodule columns
tumPercentByTumorDiamsTrim <- tumPercentByTumorDiams %>% select(-contains("largest_diameter"), -contains("%_tum_ann"), -contains("percentxdiam_annot"))

# Dichotomize sum tumor score annot
hist(tumPercentByTumorDiamsTrim$sum_tumour_score_annot)
tumPercentByTumorDiamsTrim <- tumPercentByTumorDiamsTrim %>% mutate(sum_tumour_score_annot_group = ifelse(sum_tumour_score_annot >= 250, "250+", "<250"))
tumPercentByTumorDiamsTrim <- tumPercentByTumorDiamsTrim %>% relocate(sum_tumour_score_annot_group, .after = sum_tumour_score_annot)

clinPatAnnotationData <- merge(clinPatAnnotationData, tumPercentByTumorDiamsTrim, all.x = TRUE)
```

Merge annotation derived regression data with clinPath data
```{r }
gpByTumor <- read.csv(gpAnnotationsTumorFn)
gpByTumor <- gpByTumor %>% select(-length_um)

# TODO: Check GP annotations for 89-a
# Initialize % of replacement for every tumor, so those with 0 % are captured in next filter
for(an_id in unique(gpByTumor$ids)) {
  for(a_tumor in unique(gpByTumor[ gpByTumor$ids == an_id, ]$tumors)) {
    gpByTumor <- gpByTumor %>% add_row(ids = an_id, tumors = a_tumor, annotation_types = "R", percent_gp = 0)
  }
}

# Derive % replacement per probe
replacementByTumor <- gpByTumor %>% group_by(ids, tumors) %>% filter(annotation_types == "R" | annotation_types == "R2") %>% summarise(replacement_percent = sum(percent_gp) )
replacementByTumorWide <- replacementByTumor %>% pivot_wider(names_from = tumors, values_from = replacement_percent, names_prefix = "replacement_tum_ann_")

replacementByTumorWide <- replacementByTumorWide %>% rename(id = "ids")
replacementByTumorDiams <- merge(replacementByTumorWide, tumDiameters, by = "id")


# Check needed columns exist, create otherwise
for(a_letter in tumor_letters) {
  a_replacement_tum_ann_col_name <- str_glue("replacement_tum_ann_{a_letter}")
  if(!(a_replacement_tum_ann_col_name %in% colnames(replacementByTumorDiams))) {
    replacementByTumorDiams <-replacementByTumorDiams %>% add_column({{a_replacement_tum_ann_col_name}} := NA) 
    message("Added non-existent column {a_replacement_tum_ann_col_name} to replacementByTumorDiams")
  }
}

# Set aside % tumor and largest diameter for easier visual assessment
for(a_letter in tumor_letters) {
  largest_diameter_col_name <- str_glue("tumor_{a_letter}_largest_diameter (cm)")
  replacement_tum_ann_col_name <- str_glue("replacement_tum_ann_{a_letter}")
  replacementByTumorDiams <- replacementByTumorDiams %>% relocate({{largest_diameter_col_name}}, .after = {{replacement_tum_ann_col_name}})
}

# TODO: Implement a check so that for every tumor with diameter should be a % annotation tumor
ids_incomplete_replacement_annot <- c()
# ids_incomplete_replacement_annot_a <- c()
for(a_letter in tumor_letters) {
  largest_diameter_col_name <- str_glue('tumor_{a_letter}_largest_diameter (cm)')
  replacement_annot_col_name <- str_glue('replacement_tum_ann_{a_letter}')
  
  # We know a tumor exists is diameter is not NA, obtain them
  paired_columns_annot <- replacementByTumorDiams %>% select(id, {{largest_diameter_col_name}}, {{replacement_annot_col_name}}) %>% filter(!is.na(.data[[largest_diameter_col_name]]))
  # Obtain if for any existing tumor (largest diameter != NA) percent tumor is missing (NA)
  incomplete_cases_annot <- paired_columns_annot %>% filter(is.na(.data[[replacement_annot_col_name]]))
  
  for(an_id in incomplete_cases_annot$id) {
    ids_incomplete_replacement_annot <- c(ids_incomplete_replacement_annot, an_id)
    warning(str_glue("Missing replacement percent from annotation for probe id {an_id}, tumor {a_letter} "))
    print("")
  }
}

for(a_letter in tumor_letters) {
  replacementxdiam_annot_col_name <- str_glue('replacementxdiam_annot_{a_letter}') 
  largest_diameter_col_name <- str_glue('tumor_{a_letter}_largest_diameter (cm)')
  replacement_annot_col_name <- str_glue('replacement_tum_ann_{a_letter}')
  
  replacementByTumorDiams <- replacementByTumorDiams %>% mutate({{replacementxdiam_annot_col_name}} := .data[[largest_diameter_col_name]] * .data[[replacement_annot_col_name]])
  replacementByTumorDiams <- replacementByTumorDiams %>% relocate({{replacementxdiam_annot_col_name}}, .after = {{replacement_annot_col_name}})
}

# Derive sum regression score per probe (sum of % tumor x diameter)
# TODO: sum of NA (or a product of it) is 0. We need to garantee every diameter is matched with percent tumour. Add integrity check code
replacementByTumorDiams <- replacementByTumorDiams %>% rowwise() %>% mutate(sum_replacement_score_annot = sum(c_across(starts_with("replacementxdiam_annot")), na.rm = TRUE))

replacementByTumorDiams <- replacementByTumorDiams %>% relocate(sum_replacement_score_annot, .after = id)
#replacementByTumorDiams <- replacementByTumorDiams %>% relocate(sum_replacement_score_annot_old, .after = id)

# Leave out per tumour nodule columns
replacementByTumorDiamsTrim <- replacementByTumorDiams %>% select(-contains("largest_diameter"), -contains("replacement_tum_ann"), -contains("replacementxdiam"))

# Dichotomize sum replacement score annot
hist(replacementByTumorDiamsTrim$sum_replacement_score_annot)
replacementByTumorDiamsTrim <- replacementByTumorDiamsTrim %>% mutate(sum_replacement_score_annot_group = ifelse(sum_replacement_score_annot == 0, "0",
                                                                ifelse(sum_replacement_score_annot > 0 & sum_replacement_score_annot < 150, "1-150", "150+")))
replacementByTumorDiamsTrim <- replacementByTumorDiamsTrim %>% relocate(sum_replacement_score_annot_group, .after = sum_replacement_score_annot)

clinPatAnnotationData <- merge(clinPatAnnotationData, replacementByTumorDiamsTrim, all.x = TRUE)
```

Final cleanup 
```{r }
# Optimal cutpoints using maximally selected rank statistics
# OS
if(is_visual_scores) {
surv_cutpoint(clinPatAnnotationData, time = "time_fu_os", event = "status_fu_os", variables = c("age", "num_tumors", "max_diameter", "sum_diameter", "marginal", "mean_tumor_percent_report", "sum_tumour_score_report", "replacement_percent", "desmo_percent", "tum_percent_annot", "sum_tumour_score_annot", "sum_replacement_score_annot"), progressbar = FALSE) 
} else {
  surv_cutpoint(clinPatAnnotationData, time = "time_fu_os", event = "status_fu_os", variables = c("age", "num_tumors", "max_diameter", "sum_diameter", "marginal", "mean_tumor_percent_report", "sum_tumour_score_report", "replacement_percent", "desmo_percent", "percent_r2", "tum_percent_annot", "sum_tumour_score_annot", "sum_replacement_score_annot"), progressbar = FALSE) 
}

# DFS
if(is_visual_scores) {
  surv_cutpoint(clinPatAnnotationData, time = "time_fu_prog", event = "status_fu_prog", variables = c("age", "num_tumors", "max_diameter", "sum_diameter", "marginal", "mean_tumor_percent_report", "sum_tumour_score_report", "replacement_percent", "desmo_percent", "tum_percent_annot", "sum_tumour_score_annot", "sum_replacement_score_annot"), progressbar = FALSE)
} else {
surv_cutpoint(clinPatAnnotationData, time = "time_fu_prog", event = "status_fu_prog", variables = c("age", "num_tumors", "max_diameter", "sum_diameter", "marginal", "mean_tumor_percent_report", "sum_tumour_score_report", "replacement_percent", "desmo_percent", "percent_r2", "tum_percent_annot", "sum_tumour_score_annot", "sum_replacement_score_annot"), progressbar = FALSE)  
}


clinPatAnnotationData <- clinPatAnnotationData %>% rename(age_ = age_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(num_tumors_ = num_tumors_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(max_tum_diameter = max_diameter_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(sum_tum_diameter = sum_diameter_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(marginal0 = marginal_group0)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(marginal1 = marginal_group1)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(marginal01 = marginal_group01)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(tum_percent_report = tum_percent_report_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(tum_score_report = sum_tumour_score_report_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(tum_percent_annot_ = tum_percent_annot_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(tum_score_annot_ = sum_tumour_score_annot_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(replacement = replacement_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(replacement_score = sum_replacement_score_annot_group)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(desmo95 = desmo_group95)
clinPatAnnotationData <- clinPatAnnotationData %>% rename(desmo10050 = desmo_group10050)

write.csv(clinPatAnnotationData, clinPatAnnotFn, row.names = FALSE)
```
