==> 7_survival_analysis_os_pfs.Rmd <==
---
title: "7_survival_analysis_os_pfs"
#output: html_document
output: 
  pdf_document:
    keep_md: TRUE
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Generated on: `r Sys.time()`

Import required packages
```{r}
library(tidyverse)
library(survival)
library(survminer)
library(corrplot)
library(prodlim)
```

```{r}
clinPatAnnotFn <- "./output/clin_data_annot.csv"

is_visual_scores = FALSE # TRUE   #     Use visual scores data
```

Import data
```{r}
clinPatAnnotData <- read.csv(clinPatAnnotFn)

clinPatAnnotData$gp_predo3 <- as.factor(clinPatAnnotData$gp_predo3)
clinPatAnnotData$gp_predo3 <- fct_relevel(clinPatAnnotData$gp_predo3, "R", "D", "P")

clinPatAnnotData$desmo_galjart <- as.factor(clinPatAnnotData$desmo_galjart)
clinPatAnnotData$desmo_galjart <- fct_relevel(clinPatAnnotData$desmo_galjart, "(0.1,33]", "[0,0.1]",  "(33,66]","(66,99.9]",  "(99.9,100]")

clinPatAnnotData$desmo_ks <- as.factor(clinPatAnnotData$desmo_ks)
clinPatAnnotData$desmo_ks <- fct_relevel(clinPatAnnotData$desmo_ks, "[0,33]", "(33,99.9]", "(99.9,100]")

clinPatAnnotData$desmo_ks2 <- as.factor(clinPatAnnotData$desmo_ks2)
clinPatAnnotData$desmo_ks2 <- fct_relevel(clinPatAnnotData$desmo_ks2, "[0,33]", "(33,95]", "(95,100]")
```

Parametrized survival function (to avoid code repetition)
```{r}
plotSurv <- function(grouping, fun_data, surv_type, do_facet, cohort_type = "All") {
  
params <- list(grouping = substitute(grouping), fun_data = substitute(fun_data))

if(surv_type == "OS") {
  expr <- substitute(survfit(Surv(time = time_fu_os, event = status_fu_os,) ~ grouping, data = fun_data), params)
} else if(surv_type == "Liver PFS") {
  expr <- substitute(survfit(Surv(time = time_fu_prog, event = status_fu_prog,) ~ grouping, data = fun_data), params)
}
fit_object <- eval.parent(expr)
print(fit_object)

# All  
p <- ggsurvplot(fit_object, data = fun_data, 
    title = paste(surv_type, "- ", cohort_type),
    xlab = "Years", 
    ylab = "Survival probability",
    pval = TRUE, 
    risk.table = TRUE,
    risk.table.y.text = FALSE, 
    risk.table.height = 0.3, 
    xscale = 365.25, # converts days->years at xscale
    break.time.by=365.25) # sets the axis breaks to one year
print(p)

# Neoadjuvant vs Chemonaïve
if(do_facet) {
  p <- ggsurvplot_facet(fit_object, data = fun_data, 
      title = surv_type,
      xlab = "Years", 
      ylab = "Survival probability",
      pval = TRUE, 
      facet.by = "neoadjuvant", 
      xscale = 365.25, 
      break.time.by=365.25)  
  print(p)
 }
}
```

Estimation of the median follow-up time (reverse Kaplan-Meier)
```{r}
# https://publicifsv.sund.ku.dk/~tag/Teaching/share/R-tutorials/Advanced-statistics/SurvivalAnalysis.html#org5880abf
quantile(prodlim(Hist(time_fu_os, status_fu_os)~1,data=clinPatAnnotData,reverse=TRUE))
#Median potential follow-up time of the  study was [q50] days (IQR: [q75 days;upp50 days]). This means that 50% of the patients would have been observed for at least q505 days had there been no events. 
```

Survival plots
```{r}
# Global OS All
plotSurv(1, clinPatAnnotData, "OS", FALSE)
# Global OS Neoadjuvant
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
# Global OS Chemonaïve
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
# ------------------------
# Global PFS All
plotSurv(1, clinPatAnnotData, "Liver PFS", FALSE)
# Global OS Neoadjuvant
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
# Global OS Chemonaïve
plotSurv(1, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Age
plotSurv(age_, clinPatAnnotData, "OS", TRUE)
plotSurv(age_, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(age_, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(age_, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(age_, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(age_, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Sex
plotSurv(sex, clinPatAnnotData, "OS", TRUE)
plotSurv(sex, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(sex, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(sex, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(sex, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(sex, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Primary side
plotSurv(primary_side, filter(clinPatAnnotData, primary_side != "U"), "OS", TRUE)
plotSurv(primary_side, filter(clinPatAnnotData, primary_side != "U" & neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(primary_side, filter(clinPatAnnotData, primary_side != "U" & neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(primary_side, filter(clinPatAnnotData, primary_side != "U"), "Liver PFS", TRUE)
plotSurv(primary_side, filter(clinPatAnnotData, primary_side != "U" & neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(primary_side, filter(clinPatAnnotData, primary_side != "U" & neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Sync/meta
plotSurv(sync_meta, clinPatAnnotData, "OS", TRUE)
plotSurv(sync_meta, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(sync_meta, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(sync_meta, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(sync_meta, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(sync_meta, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Neoadjuvant
plotSurv(neoadjuvant, clinPatAnnotData, "OS", FALSE)
plotSurv(neoadjuvant, clinPatAnnotData, "Liver PFS", FALSE)

# Num tumors
plotSurv(num_tumors_, clinPatAnnotData, "OS", TRUE)
plotSurv(num_tumors_, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(num_tumors_, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(num_tumors_, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(num_tumors_, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(num_tumors_, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Largest tumor diameter
plotSurv(max_tum_diameter, clinPatAnnotData, "OS", TRUE)
plotSurv(max_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(max_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(max_tum_diameter, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(max_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(max_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Sum of tumor diameters
plotSurv(sum_tum_diameter, clinPatAnnotData, "OS", TRUE)
plotSurv(sum_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(sum_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(sum_tum_diameter, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(sum_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(sum_tum_diameter, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Marginal0
plotSurv(marginal0, clinPatAnnotData, "OS", TRUE)
plotSurv(marginal0, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(marginal0, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(marginal0, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(marginal0, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(marginal0, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Marginal1
plotSurv(marginal1, clinPatAnnotData, "OS", TRUE)
plotSurv(marginal1, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(marginal1, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(marginal1, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(marginal1, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(marginal1, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Marginal1
plotSurv(marginal01, clinPatAnnotData, "OS", TRUE)
plotSurv(marginal01, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(marginal01, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(marginal01, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(marginal01, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(marginal01, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Tumor percent report
plotSurv(tum_percent_report, clinPatAnnotData, "OS", TRUE)
plotSurv(tum_percent_report, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(tum_percent_report, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(tum_percent_report, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(tum_percent_report, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(tum_percent_report, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Tumor percent annotation
plotSurv(tum_percent_annot_, clinPatAnnotData, "OS", TRUE)
plotSurv(tum_percent_annot_, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(tum_percent_annot_, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(tum_percent_annot_, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(tum_percent_annot_, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(tum_percent_annot_, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

plotSurv(tum_percent_annot_5, clinPatAnnotData, "OS", TRUE)
plotSurv(tum_percent_annot_5, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(tum_percent_annot_5, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(tum_percent_annot_5, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(tum_percent_annot_5, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(tum_percent_annot_5, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Predominant GP (D, R, P)
plotSurv(gp_predo3, clinPatAnnotData, "OS", TRUE)
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(gp_predo3, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Predominant GP (D, R1, R2, P)
plotSurv(gp_predo4, clinPatAnnotData, "OS", TRUE)
plotSurv(gp_predo4, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(gp_predo4, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(gp_predo4, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(gp_predo4, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(gp_predo4, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Predominant GP (D, R) OS All
plotSurv(gp_predo3, filter(clinPatAnnotData, gp_predo3 != 'P'), "OS", TRUE)
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(gp_predo3, filter(clinPatAnnotData, gp_predo3 != 'P'), "Liver PFS", TRUE)
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(gp_predo3, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Percent replacement
plotSurv(replacement, clinPatAnnotData, "OS", TRUE)
plotSurv(replacement, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(replacement, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(replacement, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(replacement, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(replacement, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Desmo95 cut-values
plotSurv(desmo95, clinPatAnnotData, "OS", TRUE)
plotSurv(desmo95, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(desmo95, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(desmo95, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(desmo95, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(desmo95, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Desmo10050 cut-values
plotSurv(desmo10050, clinPatAnnotData, "OS", TRUE)
plotSurv(desmo10050, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(desmo10050, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(desmo10050, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(desmo10050, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(desmo10050, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")


# Desmo Galjart values
plotSurv(desmo_galjart, clinPatAnnotData, "OS", TRUE)
plotSurv(desmo_galjart, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(desmo_galjart, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(desmo_galjart, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(desmo_galjart, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(desmo_galjart, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Desmo KS values
plotSurv(desmo_ks, clinPatAnnotData, "OS", TRUE)
plotSurv(desmo_ks, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(desmo_ks, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(desmo_ks, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(desmo_ks, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(desmo_ks, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Desmo KS values with with 95% cut-off
plotSurv(desmo_ks2, clinPatAnnotData, "OS", TRUE)
plotSurv(desmo_ks2, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(desmo_ks2, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(desmo_ks2, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(desmo_ks2, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(desmo_ks2, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Desmo Galjart values without 100 % (to see if diffs between other proportions)
#plotSurv(desmo_galjart, filter(clinPatAnnotData, desmo_galjart != "(99.9,100]"), "OS", TRUE)
#plotSurv(desmo_galjart, filter(clinPatAnnotData, desmo_galjart != "(99.9,100]"), "PFS", TRUE)
```

Experimental parameters
```{r}
# Summed tumor score report, calculated as the sum of % tumor cells x diameter
plotSurv(tum_score_report, clinPatAnnotData, "OS", TRUE)
plotSurv(tum_score_report, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(tum_score_report, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(tum_score_report, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(tum_score_report, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(tum_score_report, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Summed tumor score annotation, calculated as the sum of % tumor cells x diameter
plotSurv(tum_score_annot_, clinPatAnnotData, "OS", TRUE)
plotSurv(tum_score_annot_, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(tum_score_annot_, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(tum_score_annot_, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(tum_score_annot_, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(tum_score_annot_, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# Summed replacement score, calculated as the sum of % replacement x diameter
plotSurv(replacement_score, clinPatAnnotData, "OS", TRUE)
plotSurv(replacement_score, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(replacement_score, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(replacement_score, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(replacement_score, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(replacement_score, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")

# % Replacement 2 vs 1 (for cases with replacement)
plotSurv(r2_group, clinPatAnnotData, "OS", TRUE)
plotSurv(r2_group, filter(clinPatAnnotData, neoadjuvant == "Y"), "OS", FALSE, "Neoadjuvant")
plotSurv(r2_group, filter(clinPatAnnotData, neoadjuvant == "N"), "OS", FALSE, "Chemonaïve")
plotSurv(r2_group, clinPatAnnotData, "Liver PFS", TRUE)
plotSurv(r2_group, filter(clinPatAnnotData, neoadjuvant == "Y"), "Liver PFS", FALSE, "Neoadjuvant")
plotSurv(r2_group, filter(clinPatAnnotData, neoadjuvant == "N"), "Liver PFS", FALSE, "Chemonaïve")
```

Multivariate analysis (CCox proportional hazards model)
```{r, fig.width= 10}
# Fit a Cox proportional hazards model
# Alternatively, build model with continuous data instead of discretized

# ALL ------------------------
# UNI - OS - desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ desmo_galjart, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "UNI - OS All - desmo_galjart")

# UNI - OS - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ desmo_ks, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "UNI - OS All - desmo_ks")

# MULTI - OS desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ age_ + primary_side + num_tumors_ + max_tum_diameter + marginal1 +
                    tum_percent_annot_ + desmo_galjart, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "OS All - desmo_galjart")

# MULTI - OS desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ age_ + primary_side + num_tumors_ + max_tum_diameter + marginal1 +
                    tum_percent_annot_ + desmo_ks, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "OS All - desmo_ks")

# UNI - Liver PFS - desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ desmo_galjart, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "UNI - Liver PFS All - desmo_galjart")

# UNI - Liver PFS - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ desmo_ks, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "UNI - Liver PFS All - desmo_ks")

# MULTI - Liver PFS desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ primary_side + sync_meta + num_tumors_ + max_tum_diameter + marginal1 +
                    desmo_galjart, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "Liver PFS All - desmo_galjart")

# MULTI - Liver PFS desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ primary_side + sync_meta + num_tumors_ + max_tum_diameter + marginal1 +
                    desmo_ks, data = clinPatAnnotData)
summary(fit.coxph)
ggforest(fit.coxph, data = clinPatAnnotData, main = "Liver PFS All - desmo_ks")

# OS ------------------------
# UNI - OS Neoadjuvant - desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ desmo_galjart, data = filter(clinPatAnnotData, neoadjuvant == "Y"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y"), main = "UNI - OS Neoadjuvant - desmo_galjart")

# UNI - OS Neoadjuvant - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ desmo_ks, data = filter(clinPatAnnotData, neoadjuvant == "Y"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y"), main = "UNI - OS Neoadjuvant - desmo_KS")

## MULTI - OS Neoadjuvant - desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ age_ + num_tumors_ + max_tum_diameter + marginal1 +
                    tum_percent_annot_ + desmo_galjart, data = filter(clinPatAnnotData, neoadjuvant == "Y"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y"), main = "OS Neoadjuvant - desmo_galjart")

## MULTI - OS Neoadjuvant - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ age_ + num_tumors_ + max_tum_diameter + marginal1 +
                    tum_percent_annot_ + desmo_ks, data = filter(clinPatAnnotData, neoadjuvant == "Y"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y"), main = "OS Neoadjuvant - desmo_ks")

## MULTI - OS Chemonaive - gp_predo3
fit.coxph <- coxph(Surv(time = time_fu_os, event = status_fu_os) ~ age_ + num_tumors_ + max_tum_diameter + gp_predo3,
                   data = filter(clinPatAnnotData, neoadjuvant == "N"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "N"), main = "OS Chemonaive - gp_predo3")

# PFS ------------------------
# UNI - Liver PFS Neoadjuvant - desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ desmo_galjart, data = filter(clinPatAnnotData, neoadjuvant == "Y"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y"), main = "UNI - Liver PFS Neoadjuvant - desmo_galjart")

# UNI - Liver PFS Neoadjuvant - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ desmo_ks, data = filter(clinPatAnnotData, neoadjuvant == "Y"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y"), main = "UNI - Liver PFS Neoadjuvant - desmo_KS")

# MULTI - Liver PFS Neoadjuvant - desmo_galjart
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ sync_meta + num_tumors_ + max_tum_diameter + marginal1 + tum_percent_annot_ + desmo_galjart, data = filter(clinPatAnnotData, neoadjuvant == "Y", primary_side != "U"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y", primary_side != "U"), main = "Liver PFS Neoadjuvant - desmo_galjart")

# MULTI - Liver PFS Neoadjuvant - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ sync_meta + num_tumors_ + max_tum_diameter + marginal1 + tum_percent_annot_ + desmo_ks, data = filter(clinPatAnnotData, neoadjuvant == "Y", primary_side != "U"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "Y", primary_side != "U"), main = "Liver PFS Neoadjuvant - desmo_ks")

# MULTI - Liver PFS Chemonaïve - desmo_ks
fit.coxph <- coxph(Surv(time = time_fu_prog, event = status_fu_prog) ~ sync_meta + num_tumors_ + desmo_ks, data = filter(clinPatAnnotData, neoadjuvant == "N"))
summary(fit.coxph)
ggforest(fit.coxph, data = filter(clinPatAnnotData, neoadjuvant == "N"), main = "PFS Chemonaïve - desmo_ks")
```

Correlation matrix - All
```{r, fig.width= 10}
corrData <- clinPatAnnotData %>% select(where(is.numeric)) %>% select(-id)
if(is_visual_scores) {
  corrData <- corrData %>% select(-percent_r2)
}

corr_matrix <-cor(corrData, use = "na.or.complete")
head(round(corr_matrix,2))

corrplot(corr_matrix, type="upper", order="hclust")

# mat : is a matrix of data
# ... : further arguments to pass to the native R cor.test function
cor.mtest <- function(mat, ...) {
    mat <- as.matrix(mat)
    n <- ncol(mat)
    p.mat<- matrix(NA, n, n)
    diag(p.mat) <- 0
    for (i in 1:(n - 1)) {
        for (j in (i + 1):n) {
            tmp <- cor.test(mat[, i], mat[, j], ...)
            p.mat[i, j] <- p.mat[j, i] <- tmp$p.value
        }
    }
  colnames(p.mat) <- rownames(p.mat) <- colnames(mat)
  p.mat
}
# matrix of the p-value of the correlation
p.mat <- cor.mtest(corrData)
head(p.mat[, 1:5])

# Leave blank on no significant coefficient, 0.05
corrplot(corr_matrix, type="upper", order="hclust", 
         p.mat = p.mat, sig.level = 0.05, insig = "blank")
```

Correlation matrix - Neoadj
```{r, fig.width= 10}
corrData <- clinPatAnnotData %>% filter (neoadjuvant == "Y") %>% select(where(is.numeric)) %>% select(-id)
if(is_visual_scores) {
  corrData <- corrData %>% select(-percent_r2)
}

corr_matrix <-cor(corrData, use = "na.or.complete")

# matrix of the p-value of the correlation
p.mat <- cor.mtest(corrData)

# Leave blank on no significant coefficient, 0.05
corrplot(corr_matrix, type="upper", order="hclust", 
         p.mat = p.mat, sig.level = 0.05, insig = "blank")
```

Correlation matrix - Chemonaive
```{r, fig.width= 10}
corrData <- clinPatAnnotData %>% filter (neoadjuvant == "N") %>% select(where(is.numeric)) %>% select(-id)
if(is_visual_scores) {
  corrData <- corrData %>% select(-percent_r2)
}

corr_matrix <-cor(corrData, use = "na.or.complete")

# matrix of the p-value of the correlation
p.mat <- cor.mtest(corrData)

# Leave blank on no significant coefficient, 0.05
corrplot(corr_matrix, type="upper", order="hclust", 
         p.mat = p.mat, sig.level = 0.05, insig = "blank")
```

Interesting patient groups
```{r}
# Pure desmo according to Galjart
table(clinPatAnnotData$desmo_galjart)
pure_desmo_galjart <- filter(clinPatAnnotData, desmo_galjart == "(99.9,100]") %>% arrange(neoadjuvant)
head(pure_desmo_galjart)
table(pure_desmo_galjart$neoadjuvant)

#write.csv(pure_desmo_galjart, "./output/pure_desmo_galjart.csv")

# Desmo >= 95
desmo95 <- filter(clinPatAnnotData, desmo_percent >= 95) %>% arrange(neoadjuvant)
#write.csv(desmo95, "./output/desmo_95.csv", row.names=FALSE)

# Replacement >= 90
replacement90 <- filter(clinPatAnnotData, replacement_percent >= 90) %>% arrange(neoadjuvant)
#write.csv(replacement90, "./output/replacement_90.csv", row.names=FALSE)
```
